pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zotero-chunk-rag"
version = "0.1.0"
description = "Chunk-level semantic search for Zotero libraries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.4.0",
    "google-genai>=1.0.0",
    "pymupdf>=1.23.0",
    "fastmcp>=2.0.0",
    "tqdm>=4.0.0",
    "pydantic>=2.0.0",
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=9.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=9.0.0",
]

[project.scripts]
zotero-chunk-rag = "zotero_chunk_rag.server:mcp.run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
zotero_chunk_rag = ["data/*.csv"]
